// Resturant.Public.Web/ClientApp/src/app/auth/Models/User.ts
export interface User {
  id?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  phoneNumber?: string;
  name?: string;
  imageUrl?: string;
  status?: string;
  membershipType?: string;
  institutionId?: string;
  institutionName?: string;
  emailConfirmed?: boolean;
  roles?: string[];
}